feat!: rename rule shebang => hashbang, deprecate rule shebang
fixes #https://github.com/eslint-community/eslint-plugin-n/issues/196
