+++ lib/configs/_commons.js
@@ -16,6 +16,6 @@ module.exports = {
         "n/no-unsupported-features/es-syntax": "error",
         "n/no-unsupported-features/node-builtins": "error",
         "n/process-exit-as-throw": "error",
-        "n/shebang": "error",
+        "n/hashbang": "error",
     },
 }
