+++ lib/index.js
@@ -43,10 +43,11 @@ const rules = {
     "prefer-promises/dns": require("./rules/prefer-promises/dns"),
     "prefer-promises/fs": require("./rules/prefer-promises/fs"),
     "process-exit-as-throw": require("./rules/process-exit-as-throw"),
-    shebang: require("./rules/shebang"),
+    hashbang: require("./rules/hashbang"),
 
     // Deprecated rules.
     "no-hide-core-modules": require("./rules/no-hide-core-modules"),
+    shebang: require("./rules/shebang"),
 }
 
 const mod = {
